rename filenames use lowercase letters
