+++ applications/mne_anonymize/settingscontrollergui.h
@@ -1,6 +1,6 @@
 //=============================================================================================================
 /**
- * @file     settingscontrollerGUI.cpp
+ * @file     settingscontrollergui.h
  * @author   Juan GPC <juangpc@gmail.com>
  * @since    0.1.0
  * @date     May, 2020
@@ -28,41 +28,77 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  *
- * @brief    SettingsControllerGUI class definition.
+ * @brief     SettingsControllerGUI class declaration.
  *
  */
+
+#ifndef MNEANONYMIZE_SETTINGSCONTROLLERGUI_H
+#define MNEANONYMIZE_SETTINGSCONTROLLERGUI_H
 
 //=============================================================================================================
 // INCLUDES
 //=============================================================================================================
 
-#include "settingscontrollerGUI.h"
+#include "settingscontrollercl.h"
 
 //=============================================================================================================
 // QT INCLUDES
 //=============================================================================================================
+
+#include <QSharedPointer>
 
 //=============================================================================================================
 // EIGEN INCLUDES
 //=============================================================================================================
 
 //=============================================================================================================
-// USED NAMESPACES
-//=============================================================================================================
-
-using namespace MNEANONYMIZE;
-
-//=============================================================================================================
-// DEFINE GLOBAL METHODS
+// FORWARD DECLARATIONS
 //=============================================================================================================
 
 //=============================================================================================================
-// DEFINE MEMBER METHODS
+// DEFINE NAMESPACE MNEANONYMIZE
 //=============================================================================================================
 
-SettingsControllerGUI::SettingsControllerGUI(const QStringList& arguments)
+namespace MNEANONYMIZE {
+
+
+//=============================================================================================================
+// MNEANONYMIZE FORWARD DECLARATIONS
+//=============================================================================================================
+
+//=============================================================================================================
+/**
+ * Description of what this class is intended to do (in detail).
+ *
+ * @brief Brief description of this class.
+ */
+class SettingsControllerGUI : public SettingsControllerCL
 {
+    Q_OBJECT
 
-}
+public:
+    typedef QSharedPointer<SettingsControllerGUI> SPtr;            /**< Shared pointer type for SettingsControllerGUI. */
+    typedef QSharedPointer<const SettingsControllerGUI> ConstSPtr; /**< Const shared pointer type for SettingsControllerGUI. */
+
+    //=========================================================================================================
+    /**
+    * Constructs a SettingsControllerGUI object.
+    */
+    explicit SettingsControllerGUI(const QStringList& arguments);
+
+protected:
+
+private:
+
+signals:
+};
 
 //=============================================================================================================
+// INLINE DEFINITIONS
+//=============================================================================================================
+
+
+} // namespace MNEANONYMIZE
+
+#endif // MNEANONYMIZE_SETTINGSCONTROLLERGUI_H
+
